[ Misc ] Refactor Marlin Python Utilities (#6082)
